requirements-ui.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Existing travel agent dependencies
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.7
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.8

